add Github workflows

+++ scripts/distribute-workflow-files.sh
@@ -11,28 +11,40 @@ echo "version=${version}";
 pushd ./workflow || exit 1;
 mkdir -p ../dist || exit 1;
 
+# https://stackoverflow.com/questions/2875424/correct-way-to-check-for-a-command-line-flag-in-bash
+has_param() {
+  local term="$1";
+  shift;
+  for arg; do [ $arg == "$term" ] && return 0; done
+  return 1;
+}
+
 echo "[.] cleaning files";
 find . -type f \( -name '._*' -o -iname '.DS_Store' -o -iname '*.js.map' \) -delete || exit 1;
 
-if test -d libs; then rm -r libs || exit 1; fi
-mkdir libs || exit 1;
+if has_param "--all"; then
 
-echo "[.] packing workflow $version with arm64 Node.js runtime ...";
-cp -v ../libs/node_arm64 ./libs/ || exit 1;
-target_file="../dist/OpenInVSCode.arm64.alfredworkflow";
-test -f "$target_file" && rm "$target_file";
-zip -r "$target_file" . || exit 1;
+  if test -d libs; then rm -r libs || exit 1; fi
+  mkdir libs || exit 1;
 
-echo "[.] packing workflow $version with arm64 Node.js runtime ...";
-rm -r libs || exit 1;
-mkdir libs || exit 1;
-cp -v ../libs/node_x64 ./libs/ || exit 1;
-target_file="../dist/OpenInVSCode.amd64.alfredworkflow";
-test -f "$target_file" && rm "$target_file";
-zip -r "$target_file" . || exit 1;
+  echo "[.] packing workflow $version with arm64 Node.js runtime ...";
+  cp -v ../libs/node_arm64 ./libs/ || exit 1;
+  target_file="../dist/OpenInVSCode.arm64.alfredworkflow";
+  test -f "$target_file" && rm "$target_file";
+  zip -r "$target_file" . || exit 1;
+
+  echo "[.] packing workflow $version with arm64 Node.js runtime ...";
+  rm -r libs || exit 1;
+  mkdir libs || exit 1;
+  cp -v ../libs/node_x64 ./libs/ || exit 1;
+  target_file="../dist/OpenInVSCode.amd64.alfredworkflow";
+  test -f "$target_file" && rm "$target_file";
+  zip -r "$target_file" . || exit 1;
+
+fi
 
 echo "[.] packing workflow $version without Node.js runtime ...";
-rm -r libs || exit 1;
+if test -d libs; then rm -r libs || exit 1; fi
 target_file="../dist/OpenInVSCode.alfredworkflow";
 test -f "$target_file" && rm "$target_file";
 zip -r "$target_file" . || exit 1;
